Add files via upload
This program scrapes text from multiple FOMC statements and generates two wordclouds corresponding to time periods before and after the recent COVID recession. The wordcloud jumbles the words in a seemingly random way with different colors to give a visual impression of the texts' focus and tone. This analysis compares the focus and tone of the FOMC before and after the recession.